Separate fmt and clippy
